+++ CorrugatedIron/RiakAsyncClient.cs
@@ -14,8 +14,11 @@
 // specific language governing permissions and limitations
 // under the License.
 
+using System.Numerics;
 using CorrugatedIron.Models;
+using CorrugatedIron.Models.Index;
 using CorrugatedIron.Models.MapReduce;
+using CorrugatedIron.Models.Search;
 using CorrugatedIron.Util;
 using System;
 using System.Collections.Generic;
@@ -31,6 +34,9 @@ namespace CorrugatedIron
         Task<RiakResult<RiakObject>> Get(RiakObjectId objectId, RiakGetOptions options = null);
         Task<IEnumerable<RiakResult<RiakObject>>> Get(IEnumerable<RiakObjectId> bucketKeyPairs, RiakGetOptions options = null);
 
+        Task<RiakCounterResult> IncrementCounter(string bucket, string counter, long amount, RiakCounterUpdateOptions options = null);
+        Task<RiakCounterResult> GetCounter(string bucket, string counter, RiakCounterGetOptions options = null);
+
         Task<RiakResult<RiakObject>> Put(RiakObject value, RiakPutOptions options = null);
         Task<IEnumerable<RiakResult<RiakObject>>> Put(IEnumerable<RiakObject> values, RiakPutOptions options = null);
 
@@ -39,73 +45,84 @@ namespace CorrugatedIron
         Task<IEnumerable<RiakResult>> Delete(IEnumerable<RiakObjectId> objectIds, RiakDeleteOptions options = null);
         Task<IEnumerable<RiakResult>> DeleteBucket(string bucket, uint rwVal = RiakConstants.Defaults.RVal);
 
+        Task<RiakResult<RiakSearchResult>> Search(RiakSearchRequest search);
+
         Task<RiakResult<RiakMapReduceResult>> MapReduce(RiakMapReduceQuery query);
         Task<RiakResult<RiakStreamedMapReduceResult>> StreamMapReduce(RiakMapReduceQuery query);
 
         Task<RiakResult<IEnumerable<string>>> ListBuckets();
+        Task<RiakResult<IEnumerable<string>>> StreamListBuckets();
         Task<RiakResult<IEnumerable<string>>> ListKeys(string bucket);
         Task<RiakResult<IEnumerable<string>>> StreamListKeys(string bucket);
 
-        Task<RiakResult<RiakBucketProperties>> GetBucketProperties(string bucket, bool extended = false);
-        Task<RiakResult> SetBucketProperties(string bucket, RiakBucketProperties properties);
-
-        Task<RiakResult<IList<string>>> IndexGet(string bucket, string indexName, int value);
-        Task<RiakResult<IList<string>>> IndexGet(string bucket, string indexName, string value);
-        Task<RiakResult<IList<string>>> IndexGet(string bucket, string indexName, int minValue, int maxValue);
-        Task<RiakResult<IList<string>>> IndexGet(string bucket, string indexName, string minValue, string maxValue);
+        Task<RiakResult<RiakBucketProperties>> GetBucketProperties(string bucket);
+        Task<RiakResult> SetBucketProperties(string bucket, RiakBucketProperties properties, bool useHttp = false);
+        Task<RiakResult> ResetBucketProperties(string bucket, bool useHttp = false);
 
         Task<RiakResult<IList<RiakObject>>> WalkLinks(RiakObject riakObject, IList<RiakLink> riakLinks);
 
         Task<RiakResult<RiakServerInfo>> GetServerInfo();
 
+        Task<RiakResult<RiakIndexResult>> IndexGet(string bucket, string indexName, int value, RiakIndexGetOptions options = null);
+        Task<RiakResult<RiakIndexResult>> IndexGet(string bucket, string indexName, string value, RiakIndexGetOptions options = null);
+        Task<RiakResult<RiakIndexResult>> IndexGet(string bucket, string indexName, int minValue, int maxValue, RiakIndexGetOptions options = null);
+        Task<RiakResult<RiakIndexResult>> IndexGet(string bucket, string indexName, string minValue, string maxValue, RiakIndexGetOptions options = null);
+
+        Task<RiakResult<RiakStreamedIndexResult>> StreamIndexGet(string bucket, string indexName, BigInteger value, RiakIndexGetOptions options = null);
+        Task<RiakResult<RiakStreamedIndexResult>> StreamIndexGet(string bucket, string indexName, string value, RiakIndexGetOptions options = null);
+        Task<RiakResult<RiakStreamedIndexResult>> StreamIndexGet(string bucket, string indexName, BigInteger minValue, BigInteger maxValue, RiakIndexGetOptions options = null);
+        Task<RiakResult<RiakStreamedIndexResult>> StreamIndexGet(string bucket, string indexName, string minValue, string maxValue, RiakIndexGetOptions options = null);
+
+        Task<RiakResult<IList<string>>> ListKeysFromIndex(string bucket);
+
         Task Batch(Action<IRiakBatchClient> batchAction);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void Ping(Action<RiakResult> callback);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void Put(RiakObject value, Action<RiakResult<RiakObject>> callback, RiakPutOptions options = null);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void Put(IEnumerable<RiakObject> values, Action<IEnumerable<RiakResult<RiakObject>>> callback, RiakPutOptions options = null);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void Delete(string bucket, string key, Action<RiakResult> callback, RiakDeleteOptions options = null);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void Delete(RiakObjectId objectId, Action<RiakResult> callback, RiakDeleteOptions options = null);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void Delete(IEnumerable<RiakObjectId> objectIds, Action<IEnumerable<RiakResult>> callback, RiakDeleteOptions options = null);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void DeleteBucket(string bucket, Action<IEnumerable<RiakResult>> callback, uint rwVal = RiakConstants.Defaults.RVal);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void MapReduce(RiakMapReduceQuery query, Action<RiakResult<RiakMapReduceResult>> callback);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void StreamMapReduce(RiakMapReduceQuery query, Action<RiakResult<RiakStreamedMapReduceResult>> callback);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void ListBuckets(Action<RiakResult<IEnumerable<string>>> callback);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void ListKeys(string bucket, Action<RiakResult<IEnumerable<string>>> callback);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void StreamListKeys(string bucket, Action<RiakResult<IEnumerable<string>>> callback);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
-        void GetBucketProperties(string bucket, Action<RiakResult<RiakBucketProperties>> callback, bool extended = false);
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
+        void GetBucketProperties(string bucket, Action<RiakResult<RiakBucketProperties>> callback);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void SetBucketProperties(string bucket, RiakBucketProperties properties, Action<RiakResult> callback);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void WalkLinks(RiakObject riakObject, IList<RiakLink> riakLinks, Action<RiakResult<IList<RiakObject>>> callback);
 
-        [Obsolete("All async operations should use the functions that return Task<T>.")]
+        [Obsolete("All async operations should use the functions that return Task<T>. This will be removed in CorrugatedIron 1.5.")]
         void GetServerInfo(Action<RiakResult<RiakServerInfo>> callback);
     }
 
@@ -141,6 +158,16 @@ namespace CorrugatedIron
             return Task.Factory.StartNew(() => _client.Get(bucketKeyPairs, options));
         }
 
+        public Task<RiakCounterResult> IncrementCounter(string bucket, string counter, long amount, RiakCounterUpdateOptions options = null)
+        {
+            return Task.Factory.StartNew(() => _client.IncrementCounter(bucket, counter, amount, options));
+        }
+
+        public Task<RiakCounterResult> GetCounter(string bucket, string counter, RiakCounterGetOptions options = null) 
+        {
+            return Task.Factory.StartNew(() => _client.GetCounter(bucket, counter, options));
+        }
+
         public Task<IEnumerable<RiakResult<RiakObject>>> Put(IEnumerable<RiakObject> values, RiakPutOptions options)
         {
             return Task.Factory.StartNew(() => _client.Put(values, options));
@@ -171,6 +198,11 @@ namespace CorrugatedIron
             return Task.Factory.StartNew(() => _client.DeleteBucket(bucket, rwVal));
         }
 
+        public Task<RiakResult<RiakSearchResult>> Search(RiakSearchRequest search)
+        {
+            return Task.Factory.StartNew(() => _client.Search(search));
+        }
+
         public Task<RiakResult<RiakMapReduceResult>> MapReduce(RiakMapReduceQuery query)
         {
             return Task.Factory.StartNew(() => _client.MapReduce(query));
@@ -179,6 +211,11 @@ namespace CorrugatedIron
         public Task<RiakResult<RiakStreamedMapReduceResult>> StreamMapReduce(RiakMapReduceQuery query)
         {
             return Task.Factory.StartNew(() => _client.StreamMapReduce(query));
+        }
+
+        public Task<RiakResult<IEnumerable<string>>> StreamListBuckets()
+        {
+            return Task.Factory.StartNew(() => _client.StreamListBuckets());
         }
 
         public Task<RiakResult<IEnumerable<string>>>  ListBuckets()
@@ -196,34 +233,64 @@ namespace CorrugatedIron
             return Task.Factory.StartNew(() => _client.StreamListKeys(bucket));
         }
 
-        public Task<RiakResult<RiakBucketProperties>> GetBucketProperties(string bucket, bool extended = false)
+        public Task<RiakResult<RiakBucketProperties>> GetBucketProperties(string bucket)
         {
-            return Task.Factory.StartNew(() => _client.GetBucketProperties(bucket, extended));
+            return Task.Factory.StartNew(() => _client.GetBucketProperties(bucket));
         }
 
-        public Task<RiakResult> SetBucketProperties(string bucket, RiakBucketProperties properties)
+        public Task<RiakResult> SetBucketProperties(string bucket, RiakBucketProperties properties, bool useHttp = false)
         {
-            return Task.Factory.StartNew(() => _client.SetBucketProperties(bucket, properties));
+            return Task.Factory.StartNew(() => _client.SetBucketProperties(bucket, properties, useHttp));
         }
 
-        public Task<RiakResult<IList<string>>> IndexGet(string bucket, string indexName, int value)
+        public Task<RiakResult> ResetBucketProperties(string bucket, bool useHttp = false)
         {
-            return Task.Factory.StartNew(() => _client.IndexGet(bucket, indexName, value));
+            return Task.Factory.StartNew(() => _client.ResetBucketProperties(bucket, useHttp));
         }
 
-        public Task<RiakResult<IList<string>>> IndexGet(string bucket, string indexName, string value)
+        public Task<RiakResult<RiakIndexResult>> IndexGet(string bucket, string indexName, int value, RiakIndexGetOptions options = null)
         {
-            return Task.Factory.StartNew(() => _client.IndexGet(bucket, indexName, value));
+            return Task.Factory.StartNew(() => _client.IndexGet(bucket, indexName, value, options));
         }
 
-        public Task<RiakResult<IList<string>>> IndexGet(string bucket, string indexName, int minValue, int maxValue)
+        public Task<RiakResult<RiakIndexResult>> IndexGet(string bucket, string indexName, string value, RiakIndexGetOptions options = null)
         {
-            return Task.Factory.StartNew(() => _client.IndexGet(bucket, indexName, minValue, maxValue));
+            return Task.Factory.StartNew(() => _client.IndexGet(bucket, indexName, value, options));
         }
 
-        public Task<RiakResult<IList<string>>> IndexGet(string bucket, string indexName, string minValue, string maxValue)
+        public Task<RiakResult<RiakIndexResult>> IndexGet(string bucket, string indexName, int minValue, int maxValue, RiakIndexGetOptions options = null)
         {
-            return Task.Factory.StartNew(() => _client.IndexGet(bucket, indexName, minValue, maxValue));
+            return Task.Factory.StartNew(() => _client.IndexGet(bucket, indexName, minValue, maxValue, options));
+        }
+
+        public Task<RiakResult<RiakIndexResult>> IndexGet(string bucket, string indexName, string minValue, string maxValue, RiakIndexGetOptions options = null)
+        {
+            return Task.Factory.StartNew(() => _client.IndexGet(bucket, indexName, minValue, maxValue, options));
+        }
+
+        public Task<RiakResult<RiakStreamedIndexResult>> StreamIndexGet(string bucket, string indexName, BigInteger value, RiakIndexGetOptions options = null)
+        {
+            return Task.Factory.StartNew(() => _client.StreamIndexGet(bucket, indexName, value, options));
+        }
+
+        public Task<RiakResult<RiakStreamedIndexResult>> StreamIndexGet(string bucket, string indexName, string value, RiakIndexGetOptions options = null)
+        {
+            return Task.Factory.StartNew(() => _client.StreamIndexGet(bucket, indexName, value, options));
+        }
+
+        public Task<RiakResult<RiakStreamedIndexResult>> StreamIndexGet(string bucket, string indexName, BigInteger minValue, BigInteger maxValue, RiakIndexGetOptions options = null)
+        {
+            return Task.Factory.StartNew(() => _client.StreamIndexGet(bucket, indexName, minValue, maxValue, options));
+        }
+
+        public Task<RiakResult<RiakStreamedIndexResult>> StreamIndexGet(string bucket, string indexName, string minValue, string maxValue, RiakIndexGetOptions options = null)
+        {
+            return Task.Factory.StartNew(() => _client.StreamIndexGet(bucket, indexName, minValue, maxValue, options));
+        }
+
+        public Task<RiakResult<IList<string>>> ListKeysFromIndex(string bucket)
+        {
+            return Task.Factory.StartNew(() => _client.ListKeysFromIndex(bucket));
         }
 
         public Task<RiakResult<IList<RiakObject>>> WalkLinks(RiakObject riakObject, IList<RiakLink> riakLinks)
@@ -329,9 +396,9 @@ namespace CorrugatedIron
             ExecAsync(() => callback(_client.StreamListKeys(bucket)));
         }
 
-        public void GetBucketProperties(string bucket, Action<RiakResult<RiakBucketProperties>> callback, bool extended = false)
+        public void GetBucketProperties(string bucket, Action<RiakResult<RiakBucketProperties>> callback)
         {
-            ExecAsync(() => callback(_client.GetBucketProperties(bucket, extended)));
+            ExecAsync(() => callback(_client.GetBucketProperties(bucket)));
         }
 
         public void SetBucketProperties(string bucket, RiakBucketProperties properties, Action<RiakResult> callback)