+++ RELEASE_NOTES.md
@@ -1,6 +1,20 @@
 Release Notes
 =============
 
+v1.4.0
+------
+
+* **Support for Riak 1.4**
+* Bucket properties calls have been moved entirely to the Protocol Buffers interface. Riak 1.3 and earlier users should make sure to set the `useHttp` parameter to `true`. This won't make any difference for setting bucket properties - `extended` was renamed to `useHttp`.
+* Index pagination (#127)
+* Reseting bucket properties through PB API (#126) - no API change here, we just do it the official way.
+* Counters are available via `RiakClient.IncrementCounter` and `RiakClient.GetCounter`. Various counter options are supplied via `RiakCounterUpdateOptions` and `RiakCounterGetOptions`. (#125)
+* Removed `Either<uint, string>` in many Riak*Options classes. Users can still set options quorum options using a string or an integer, but the internal representation has changed. New code should use `RiakConstants.QuorumOptions` for setting and comparing quorum options.
+* Created `RiakIndexResult` to deal with new possible riak index results containing either a list of keys or list of key + term pairs. (#142, #128) **N.B.** this will break CI 1.3.x and earlier range query operations. 
+    * Existing queries will need to use `results.Value.IndexKeyTerms` to enumerate through the results of an 2i query.
+    * An `IndexKeyTerm` contains a `key` and `term`. `term` will be empty unless the `return_terms` query option is used.
+
+
 v1.3.3
 ------